+++ Java+/App+/GptUi/src/gptui/gpt/openai/GptApi.java
@@ -1,7 +1,5 @@
 package gptui.gpt.openai;
 
-import java.math.BigDecimal;
-
 public interface GptApi {
-    String send(String content, BigDecimal temperature);
+    String send(String content, Integer temperature);
 }
